feat(mamba3_tracker): v15/v16/v17 — DINO encoder evolution
Bundles the encoder-side experiments from the last two days.

v15 = "bump the DINO input resolution to 896"
  * `configs/v15.yaml`. Dataset image_size 448→896, DINOv2 input 448→896,
    level_sizes [32]→[64]. Each DINO patch covers half as much of the
    original scene. Adds `interpolate_pos_encoding=True` so DINOv2 can
    handle the non-default 896×896 input.
  * Result: drivetrack motion ratio settled ~45–55% (similar to v14's
    60–75% peak). Pstudio bouncy 15–45% — modest improvement on small
    objects. Val loss ~0.04. 30 k steps, 4 h 49 min wall (resumed after
    overnight host suspend; see ~/.claude/CLAUDE.md "triangulate the
    evidence" rule added in dot_claude commit db2b8a4).

v16 = "fuse multiple DINOv2 layers" (DPT-style semantic pyramid)
  * `configs/v16.yaml` + `DINOv2Encoder.fuse_layers` parameter
    (`src/mamba3_tracker/model/dino_encoder.py`). Reads blocks
    {2, 5, 8, 11} of DINOv2-small, LayerNorms each (per-layer LN —
    different layers have very different output magnitudes),
    concatenates (4·384 = 1536), projects back to 384 with a learnable
    Linear. Trainable params 1.20 M → 1.79 M.
  * Required propagator fix: `_sample_query_features` in
    `src/mamba3_tracker/model/propagator.py` now casts grid_sample's
    output to `out.dtype` before the index-put. v14/v15's whole encoder
    ran under `@torch.no_grad()` so autocast/dtype interactions were
    different; v16 needs gradients through `fuse_proj`, narrowing the
    no_grad scope, which exposed an existing dtype mismatch
    (`grid_sample` upcasts bf16 → fp32 under autocast in
    gradient-tracked contexts).
  * Result: mid-training pstudio peaks > 100% (vs v15's ~50% range) —
    promising signal that early-layer detail helps small-object
    localisation. But late-training motion ratios converged to similar
    levels as v15 (drivetrack 40–60%, pstudio 14–37%). Val loss ~0.05.
    The aggregate metric win didn't survive late training.

v17 = "drop-in DINOv2 → DINOv3 swap" (training started, in progress)
  * `configs/v17.yaml` swaps to `facebook/dinov3-vits16-pretrain-lvd1689m`.
    Same hidden_size=384 → no projection. patch=16 instead of 14, so
    grid at 896 is 56×56 (was 64×64). No layer fusion (clean v15→v17
    comparison).
  * Required encoder fix: DINOv3 prepends register tokens (Darcet et al.
    2024) to the patch sequence: `[CLS, REG_1..REG_N, patch_1..patch_P]`.
    `DINOv2Encoder` now reads `num_register_tokens` from config and
    slices off `1 + n_reg` prefix tokens (was hard-coded `[:, 1:, :]`).
    DINOv2 has 0 register tokens so behaviour is unchanged for v14–v16.

Common to all three runs (architecture + loss + data unchanged from v12):
  * Per-anchor cumsum trajectory reconstruction.
  * Scale-normalised squared-L2 in 3D + 2D + visibility Binary Cross-Entropy.
  * Official TAPVid-3D minival, 40/5/5 stratified split per subset.
  * In-training motion-ratio diagnostic (v13+) logs every val_every.

Output-dir naming convention (`memory/feedback_output_dir_naming.md`):
  * Datetime appears EXACTLY ONCE in the path. Child dirs use
    descriptive single-purpose names (`viz_ckpt30k`, `viz_step12500`,
    `eval`, `plots`) — never a second timestamp. Anti-pattern
    `outputs/track_v16_<dt>/viz_step12500_<dt>/` flagged explicitly.

`src/mamba3_tracker/train/config.py` accepts v15/v16/v17.

+++ src/mamba3_tracker/model/propagator.py
@@ -72,11 +72,17 @@ def _sample_query_features(
                 continue
             feat = pyramid_finest[b:b + 1, t_val]
             grid = grid_xy[b, sel].view(1, -1, 1, 2)
+            # `grid_sample` may upcast to fp32 under autocast in certain
+            # gradient-tracked contexts (v16 fusion triggers this where
+            # v14/v15 didn't, because v14/v15 wrapped the whole encoder in
+            # @torch.no_grad() while v16 needs gradient through fuse_proj).
+            # Cast grid to feat's dtype, then cast sampled back to out's
+            # dtype before the index_put so dtypes always line up.
             sampled = F.grid_sample(
-                feat, grid, mode="bilinear",
+                feat, grid.to(feat.dtype), mode="bilinear",
                 padding_mode="border", align_corners=True,
             )
-            out[b, sel] = sampled.squeeze(-1).squeeze(0).transpose(0, 1)
+            out[b, sel] = sampled.squeeze(-1).squeeze(0).transpose(0, 1).to(out.dtype)
     return out
